main.py
#!/usr/bin/env python3

import os
import click
from dotenv import load_dotenv
from rich.console import Console
from rich.prompt import Prompt

# Load environment variables
load_dotenv()

# Initialize rich console
console = Console()

class PromptToApp:
    def __init__(self):
        self.api_key = os.getenv('OPENAI_API_KEY')
        if not self.api_key:
            console.print("[red]Error: OPENAI_API_KEY not found in environment variables[/red]")
            raise ValueError("OPENAI_API_KEY is required")

    def process_prompt(self, prompt: str) -> str:
        """
        Process the user's prompt and generate appropriate code
        """
        # TODO: Implement OpenAI API integration
        return "Code generation will be implemented here"

    def create_project(self, project_name: str, project_type: str):
        """
        Create a new project with the specified name and type
        """
        console.print(f"[green]Creating {project_type} project: {project_name}[/green]")
        # TODO: Implement project creation logic

@click.group()
def cli():
    """Prompt-to-App AI Coding Agent"""
    pass

@cli.command()
@click.argument('prompt')
def generate(prompt):
    """Generate code based on the provided prompt"""
    agent = PromptToApp()
    try:
        code = agent.process_prompt(prompt)
        console.print("[green]Generated code:[/green]")
        console.print(code)
    except Exception as e:
        console.print(f"[red]Error: {str(e)}[/red]")

@cli.command()
@click.argument('project_name')
@click.option('--type', '-t', default='web', help='Project type (web, mobile, desktop)')
def create(project_name, type):
    """Create a new project"""
    agent = PromptToApp()
    try:
        agent.create_project(project_name, type)
    except Exception as e:
        console.print(f"[red]Error: {str(e)}[/red]")

if __name__ == '__main__':
    cli() 