requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
click>=8.0.0
rich>=10.0.0
pyyaml>=6.0.0
jinja2>=3.0.0 